Create README.md
This is my first repository with questions that I personally would like to understand before jumping off into something I honestly have zero experience in. Hopefully these are the right questions for a beginner that can help anyone understand what github is used for, how to use and apply git hub, and what equipment or specs a re required to use github. If i can learn this I am confident anyone can. I am not the brightest crayon in the box or the sharpest tool in the shed but, I have hunger to fill and crave knowledge to feed that hunger. Who knows maybe I willbpublish a github for dummies one day. Or In my case a github for dumbest. We'll see. I'm actually pretty smat i mean smart right? i dunno oh well, hope this helps.